refactor: candle feature flag
