+++ crates/beet_ml/Cargo.toml
@@ -15,6 +15,12 @@ categories.workspace = true
 # default = ["cuda"]
 # requires cuda to be installed https://developer.nvidia.com/cuda-downloads?target_os=Linux&target_arch=x86_64&Distribution=Ubuntu&target_version=24.04&target_type=deb_local
 cuda = ["candle-transformers/cuda"]
+candle = [
+	"dep:candle-core",
+	"dep:candle-transformers",
+	"dep:candle-nn",
+	"dep:tokenizers",
+]
 spatial = ["dep:beet_spatial"]
 bevy_default = ["bevy/default"]
 
@@ -36,10 +42,10 @@ strum.workspace = true
 strum_macros.workspace = true
 
 #💡 huggingface
-candle-core = "0.8"
-candle-transformers = "0.8"
-candle-nn = "0.8"
-tokenizers = { version = "0.19.1", default-features = false, features = [
+candle-core = { version = "0.8", optional = true }
+candle-transformers = { version = "0.8", optional = true }
+candle-nn = { version = "0.8", optional = true }
+tokenizers = { version = "0.19.1", optional = true, default-features = false, features = [
 	"unstable_wasm",
 ] }
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
