+++ crates/beet_ml/src/language/mod.rs
@@ -1,31 +1,39 @@
+#[cfg(feature = "candle")]
 mod bert;
-mod run_with_user_sentence;
+#[cfg(feature = "candle")]
 pub use self::bert::*;
+mod run_with_user_sentence;
 pub use self::run_with_user_sentence::*;
-mod bert_config;
-pub use self::bert_config::*;
-mod bert_loader;
-pub use self::bert_loader::*;
-mod nearest_sentence;
-pub use self::nearest_sentence::*;
-mod sentence_embeddings;
-pub use self::sentence_embeddings::*;
-
 
 use beet_flow::prelude::*;
 use bevy::prelude::*;
+use std::borrow::Cow;
 
 #[derive(Default)]
 pub struct LanguagePlugin;
 
 impl Plugin for LanguagePlugin {
 	fn build(&self, app: &mut App) {
-		app.init_asset::<Bert>()
-			.init_asset_loader::<BertLoader>()
-			.register_type::<Sentence>()
+		app.register_type::<Sentence>()
 			.add_observer(run_with_user_sentence::<()>);
+
+		#[cfg(feature = "candle")]
+		app.init_asset::<Bert>().init_asset_loader::<BertLoader>();
 
 		let world = app.world_mut();
 		world.register_component::<Sentence>();
 	}
+}
+
+
+/// This component is for use with [`SentenceFlow`]. Add to either the agent or a child behavior.
+#[derive(Debug, Clone, Component, PartialEq, Reflect)]
+#[reflect(Component)]
+pub struct Sentence(pub Cow<'static, str>);
+impl Sentence {
+	pub fn new(s: impl Into<Cow<'static, str>>) -> Self { Self(s.into()) }
+}
+
+impl Default for Sentence {
+	fn default() -> Self { Self::new("placeholder") }
 }
