+++ crates/beet_ml/src/lib.rs
@@ -6,6 +6,8 @@
 // #![deny(missing_docs)]
 #![feature(let_chains, generic_const_exprs, const_trait_impl)]
 
+mod utils;
+
 #[cfg(feature = "bevy_default")]
 pub mod frozen_lake;
 pub mod language;
@@ -13,8 +15,6 @@ pub mod language;
 pub mod rl;
 #[cfg(feature = "bevy_default")]
 pub mod rl_realtime;
-#[cfg(feature = "spatial")]
-pub mod spatial;
 #[cfg(test)]
 pub mod test_utils;
 #[cfg(target_arch = "wasm32")]
@@ -28,10 +28,9 @@ pub mod prelude {
 	pub use crate::rl::*;
 	#[cfg(feature = "bevy_default")]
 	pub use crate::rl_realtime::*;
-	#[cfg(feature = "spatial")]
-	pub use crate::spatial::*;
 	#[cfg(test)]
 	pub use crate::test_utils::*;
+	pub use crate::utils::*;
 	#[cfg(target_arch = "wasm32")]
 	pub use crate::wasm::*;
 }
