--- crates/beet_ml/src/spatial/mod.rs
@@ -1,3 +1,0 @@
-pub mod sentence_steer_target;
-#[allow(unused_imports)]
-pub use self::sentence_steer_target::*;
