+++ justfile
@@ -123,9 +123,10 @@ hello-world:
 test-ci *args:
 	cargo fmt 				--check
 	just leptosfmt 		--check
-	RUST_MIN_STACK=16777216 cargo test --workspace					--features=_doctest {{args}} -- --test-threads=8
-	RUST_MIN_STACK=16777216 cargo test --workspace --lib 	--all-features				{{args}} -- --test-threads=8
-	cargo test --target wasm32-unknown-unknown 	--all-features	-p beet_flow 		{{args}} -- --test-threads=8
+	RUST_MIN_STACK=16777216 cargo test --workspace --lib	--features=_doctest 			{{args}} -- --test-threads=8
+	RUST_MIN_STACK=16777216 cargo test --workspace --doc	--features=_doctest 			{{args}} -- --test-threads=8
+	RUST_MIN_STACK=16777216 cargo test --workspace --lib 	--all-features						{{args}} -- --test-threads=8
+	cargo test --target wasm32-unknown-unknown 	--all-features	-p beet_flow 				{{args}} -- --test-threads=8
 
 # rebuilding bevy_render for wasm results in 'no space left on device'
 test-all *args:
